time tests are ready

+++ Task_2_1_1/src/main/java/ru/nsu/fitkulin/ParallelStreamThread.java
@@ -3,11 +3,11 @@ package ru.nsu.fitkulin;
 import java.util.Arrays;
 
 public class ParallelStreamThread {
-    public static boolean isPrime(int number) {
+    public static boolean isPrime(long number) {
         if (number < 2) {
             return false;
         }
-        for (int i = 2; i * i <= number; i++) {
+        for (long i = 2; i * i <= number; i++) {
             if (number % i == 0) {
                 return false;
             }
@@ -15,7 +15,7 @@ public class ParallelStreamThread {
         return true;
     }
 
-    public static boolean containsNonPrimeParallelStream(int[] numbers) {
+    public static boolean containsNonPrimeParallelStream(long[] numbers) {
         return Arrays.stream(numbers).parallel().anyMatch(num -> !isPrime(num));
     }
 }
